Subscribe users to model metered plans when they use the model
